+++ src/SymfonyCache/EventDispatchingHttpCache.php
@@ -109,8 +109,8 @@ trait EventDispatchingHttpCache
     {
         $response = $this->dispatch(Events::PRE_STORE, $request, $response);
 
-        // CustomTtlListener or other Listener or Subscribers might have set a non-cacheable response so we need revalidate this
-        // So store is only called when cacheable like Symfony core would do: https://github.com/symfony/symfony/blob/v7.1.2/src/Symfony/Component/HttpKernel/HttpCache/HttpCache.php#L409
+        // CustomTtlListener or other Listener or Subscribers might have changed the response to become non-cacheable, revalidate.
+        // Only call store for a cacheable response like Symfony core does: https://github.com/symfony/symfony/blob/v7.1.2/src/Symfony/Component/HttpKernel/HttpCache/HttpCache.php#L409
         if ($response->isCacheable()) {
             parent::store($request, $response);
         }